+++ shipping.sh
@@ -3,40 +3,42 @@ script_path=$(dirname $script)
 source ${script_path}/common.sh
 mysql_pwd=$1
 
+if [-z "$mysql_pwd"]; then
+  echo mysql pwd missing
+  exit
+fi
 
-echo -e "\e[31m>>>>>>>>>>>>>>Install Maven<<<<<<<<<<<<<<\e[0m"
+print_head "Install Maven"
 yum install maven -y
 
-echo -e "\e[31m>>>>>>>>>>>>>>Add app user<<<<<<<<<<<<<<\e[0m"
+print_head "Add app user"
 useradd roboshop
 
-echo -e "\e[31m>>>>>>>>>>>>>>setup an app directory<<<<<<<<<<<<<<\e[0m"
+print_head "setup an app directory"
 rm -rf /app
-echo -e "\e[31mRemove existing app directory\e[0m"
 mkdir /app
-echo -e "\e[31mOk\e[0m"
 
-echo -e "\e[31m>>>>>>>>>>>>>>Download app code<<<<<<<<<<<<<<\e[0m"
+print_head "Download app code"
 curl -L -o /tmp/shipping.zip https://roboshop-artifacts.s3.amazonaws.com/shipping.zip
 cd /app
 
-echo -e "\e[31m>>>>>>>>>>>>>>Unzip app code<<<<<<<<<<<<<<\e[0m"
+print_head "Unzip app code"
 unzip /tmp/shipping.zip
 
-echo -e "\e[31m>>>>>>>>>>>>>>Download Maven dependencies<<<<<<<<<<<<<<\e[0m"
+print_head "Download Maven dependencies"
 mvn clean package
 mv target/shipping-1.0.jar shipping.jar
 
-echo -e "\e[31m>>>>>>>>>>>>>>Copy service file<<<<<<<<<<<<<<\e[0m"
+print_head "Copy service file"
 cp ${script_path}/shipping.service /etc/systemd/system/shipping.service
 
-echo -e "\e[31m>>>>>>>>>>>>>>Install MYSQL Client<<<<<<<<<<<<<<\e[0m"
+print_head "Install MYSQL Client"
 yum install mysql -y
 
-echo -e "\e[31m>>>>>>>>>>>>>>load schema to the DB.<<<<<<<<<<<<<<\e[0m"
+print_head "load schema to the DB"
 mysql -h mysql-dev.smitdevops.online -uroot -p${mysql_pwd} < /app/schema/shipping.sql
 
-echo -e "\e[31m>>>>>>>>>>>>>>Load the service<<<<<<<<<<<<<<\e[0m"
+print_head "Load the service"
 systemctl daemon-reload
 systemctl enable shipping
 systemctl restart shipping
